Split bot utils
